column-edit PR code review first round
Update reviewTableSchema api route to camel case

Generic constants move to helpers package

Remove duplicate createDatabase function in utilities package

Rename updateTableSchema to table

Add license in table's package new files

Clean up debugger log in web2 file

Rename new file names to snake_case

Update Test names to upper camel case

Merge interleave and non-interleave test cases into a single file

Remove redundant comment for table package

Refactor remove column file

Refactor rename column file

Refactor rename and remove column files

Refactor review rename column file

Update go package
