+++ webv2/routes.go
@@ -19,7 +19,7 @@ import (
 	"github.com/cloudspannerecosystem/harbourbridge/webv2/primarykey"
 	"github.com/cloudspannerecosystem/harbourbridge/webv2/session"
 	"github.com/cloudspannerecosystem/harbourbridge/webv2/summary"
-	"github.com/cloudspannerecosystem/harbourbridge/webv2/updateTableSchema"
+	"github.com/cloudspannerecosystem/harbourbridge/webv2/table"
 
 	"github.com/gorilla/mux"
 )
@@ -37,8 +37,8 @@ func getRoutes() *mux.Router {
 	router.HandleFunc("/report", getReportFile).Methods("GET")
 	router.HandleFunc("/schema", getSchemaFile).Methods("GET")
 	router.HandleFunc("/typemap/global", setTypeMapGlobal).Methods("POST")
-	router.HandleFunc("/typemap/table", updateTableSchema.UpdateTableSchema).Methods("POST")
-	router.HandleFunc("/typemap/reviewtableschema", updateTableSchema.ReviewTableSchema).Methods("POST")
+	router.HandleFunc("/typemap/table", table.UpdateTableSchema).Methods("POST")
+	router.HandleFunc("/typemap/reviewTableSchema", table.ReviewTableSchema).Methods("POST")
 
 	router.HandleFunc("/setparent", setParentTable).Methods("GET")
