+++ webv2/session/utility.go
@@ -1,4 +1,4 @@
-package updatesessionfiles
+package session
 
 import (
 	"fmt"
@@ -6,13 +6,12 @@ import (
 
 	"github.com/cloudspannerecosystem/harbourbridge/common/utils"
 	"github.com/cloudspannerecosystem/harbourbridge/conversion"
-	"github.com/cloudspannerecosystem/harbourbridge/webv2/session"
 )
 
 // UpdateSessionFile updates the content of session file with
 // latest sessionState.Conv while also dumping schemas and report.
 func UpdateSessionFile() error {
-	sessionState := session.GetSessionState()
+	sessionState := GetSessionState()
 
 	ioHelper := &utils.IOStreams{In: os.Stdin, Out: os.Stdout}
 	_, err := conversion.WriteConvGeneratedFiles(sessionState.Conv, sessionState.DbName, sessionState.Driver, ioHelper.BytesRead, ioHelper.Out)
