+++ webv2/table/update_table_schema_test.go
@@ -1,4 +1,18 @@
-package updateTableSchema
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package table
 
 import (
 	"encoding/json"
@@ -8,16 +22,16 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/bmizerany/assert"
 	"github.com/cloudspannerecosystem/harbourbridge/common/constants"
 	"github.com/cloudspannerecosystem/harbourbridge/internal"
 	"github.com/cloudspannerecosystem/harbourbridge/proto/migration"
 	"github.com/cloudspannerecosystem/harbourbridge/schema"
 	"github.com/cloudspannerecosystem/harbourbridge/spanner/ddl"
 	"github.com/cloudspannerecosystem/harbourbridge/webv2/session"
+	"github.com/stretchr/testify/assert"
 )
 
-func TestUpdateTableSchemaV2(t *testing.T) {
+func TestUpdateTableSchema(t *testing.T) {
 
 	tc := []struct {
 		name         string
@@ -174,7 +188,7 @@ func TestUpdateTableSchemaV2(t *testing.T) {
 	}
 }
 
-func TestAddUpdateTableSchemaV2(t *testing.T) {
+func TestAddUpdateTableSchema(t *testing.T) {
 
 	tc := []struct {
 		name         string
@@ -302,7 +316,7 @@ func TestAddUpdateTableSchemaV2(t *testing.T) {
 
 }
 
-func TestChangetypeUpdateTableSchemaV2(t *testing.T) {
+func TestChangeTypeUpdateTableSchema(t *testing.T) {
 
 	tc := []struct {
 		name         string
@@ -417,6 +431,133 @@ func TestChangetypeUpdateTableSchemaV2(t *testing.T) {
 		json.Unmarshal(rr.Body.Bytes(), &res)
 
 		log.Println("res :", res.SpSchema["t1"])
+
+		if status := rr.Code; int64(status) != tc.statusCode {
+			t.Errorf("handler returned wrong status code: got %v want %v",
+				status, tc.statusCode)
+		}
+
+		if tc.statusCode == http.StatusOK {
+			assert.Equal(t, tc.expectedConv, res)
+		}
+	}
+}
+
+func TestUpdateTableSchemaInterleaved(t *testing.T) {
+
+	tc := []struct {
+		name         string
+		table        string
+		payload      string
+		statusCode   int64
+		conv         *internal.Conv
+		expectedConv *internal.Conv
+	}{
+
+		{
+			name:  "Test rename success",
+			table: "t1",
+			payload: `
+    {
+      "UpdateCols":{
+		"a": { "Rename": "aa" }
+	}
+    }`,
+			statusCode: http.StatusOK,
+			conv: &internal.Conv{
+				SpSchema: map[string]ddl.CreateTable{
+					"t1": {
+						Name:     "t1",
+						ColNames: []string{"a", "b", "c"},
+						ColDefs: map[string]ddl.ColumnDef{
+							"a": {Name: "a", T: ddl.Type{Name: ddl.Int64}, NotNull: true},
+							"b": {Name: "b", T: ddl.Type{Name: ddl.Int64}, NotNull: true},
+							"c": {Name: "c", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, NotNull: true},
+						},
+						Pks:    []ddl.IndexKey{{Col: "a", Desc: false}, {Col: "b", Desc: false}},
+						Parent: "t2",
+					},
+					"t2": {
+						Name:     "t2",
+						ColNames: []string{"a", "b", "c"},
+						ColDefs: map[string]ddl.ColumnDef{
+							"a": {Name: "a", T: ddl.Type{Name: ddl.Int64}, NotNull: true},
+							"b": {Name: "b", T: ddl.Type{Name: ddl.Int64}, NotNull: true},
+							"c": {Name: "c", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, NotNull: true},
+						},
+						Pks: []ddl.IndexKey{{Col: "a", Desc: false}},
+					},
+				},
+				ToSource: map[string]internal.NameAndCols{
+					"t1": {Name: "t1", Cols: map[string]string{"a": "a", "b": "b", "c": "c"}},
+					"t2": {Name: "t2", Cols: map[string]string{"a": "a", "b": "b", "c": "c"}},
+				},
+				ToSpanner: map[string]internal.NameAndCols{
+					"t1": {Name: "t2", Cols: map[string]string{"a": "a", "b": "b", "c": "c"}},
+					"t2": {Name: "t2", Cols: map[string]string{"a": "a", "b": "b", "c": "c"}},
+				},
+				Audit: internal.Audit{MigrationType: migration.MigrationData_SCHEMA_AND_DATA.Enum()},
+			},
+			expectedConv: &internal.Conv{
+				SpSchema: map[string]ddl.CreateTable{
+					"t1": {
+						Name:     "t1",
+						ColNames: []string{"aa", "b", "c"},
+						ColDefs: map[string]ddl.ColumnDef{
+							"aa": {Name: "aa", T: ddl.Type{Name: ddl.Int64}, NotNull: true},
+							"b":  {Name: "b", T: ddl.Type{Name: ddl.Int64}, NotNull: true},
+							"c":  {Name: "c", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, NotNull: true},
+						},
+						Pks:    []ddl.IndexKey{{Col: "aa", Desc: false}, {Col: "b", Desc: false}},
+						Parent: "t2",
+					},
+					"t2": {
+						Name:     "t2",
+						ColNames: []string{"aa", "b", "c"},
+						ColDefs: map[string]ddl.ColumnDef{
+							"aa": {Name: "aa", T: ddl.Type{Name: ddl.Int64}, NotNull: true},
+							"b":  {Name: "b", T: ddl.Type{Name: ddl.Int64}, NotNull: true},
+							"c":  {Name: "c", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, NotNull: true},
+						},
+						Pks: []ddl.IndexKey{{Col: "aa", Desc: false}},
+					},
+				},
+				ToSource: map[string]internal.NameAndCols{
+					"t1": {Name: "t1", Cols: map[string]string{"aa": "a", "b": "b", "c": "c"}},
+					"t2": {Name: "t2", Cols: map[string]string{"aa": "a", "b": "b", "c": "c"}},
+				},
+				ToSpanner: map[string]internal.NameAndCols{
+					"t1": {Name: "t2", Cols: map[string]string{"a": "aa", "b": "b", "c": "c"}},
+					"t2": {Name: "t2", Cols: map[string]string{"a": "aa", "b": "b", "c": "c"}},
+				},
+			},
+		},
+	}
+
+	for _, tc := range tc {
+
+		sessionState := session.GetSessionState()
+		sessionState.Conv = tc.conv
+		sessionState.Driver = constants.MYSQL
+
+		payload := tc.payload
+
+		req, err := http.NewRequest("POST", "/typemap/table?table="+tc.table, strings.NewReader(payload))
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		req.Header.Set("Content-Type", "application/json")
+
+		rr := httptest.NewRecorder()
+
+		handler := http.HandlerFunc(UpdateTableSchema)
+
+		handler.ServeHTTP(rr, req)
+
+		res := &internal.Conv{}
+
+		json.Unmarshal(rr.Body.Bytes(), &res)
 
 		if status := rr.Code; int64(status) != tc.statusCode {
 			t.Errorf("handler returned wrong status code: got %v want %v",
